+++ ovp_projects/serializers/project.py
@@ -199,7 +199,7 @@ class ProjectSearchSerializer(serializers.ModelSerializer):
 
   class Meta:
     model = models.Project
-    fields = ['slug', 'image', 'name', 'description', 'highlighted', 'published_date', 'address', 'organization', 'owner', 'applied_count', 'hidden_address']
+    fields = ['slug', 'image', 'name', 'description', 'highlighted', 'published_date', 'address', 'organization', 'owner', 'applied_count', 'max_applies', 'hidden_address']
 
   @hide_address
   def to_representation(self, instance):